some decent clean-up

+++ mlscorecheck/core/_rounding.py
@@ -21,6 +21,6 @@ def round_scores(to_round, rounding_decimals=None):
         return {**to_round}
 
     if not isinstance(to_round, dict):
-        return np.round(to_round, rounding_decimals)
+        return float(np.round(to_round, rounding_decimals))
 
-    return {key: np.round(value, rounding_decimals) for key, value in to_round.items()}
+    return {key: float(np.round(value, rounding_decimals)) for key, value in to_round.items()}
